Now websocket is working. Bid delete implemented in ajax.
In this commit is included:
- New artifact xml descriptor for ejbs,
- Ajax implementation of bid making and deleting
- Type conversion corrections (int -> long)

TODO:
- implement table construction and reshaping after making a bid/receiving a websocket update

+++ DistributedAuctionSystem/Interfaces/src/main/java/it/unipi/dsmt/das/ejbs/beans/interfaces/AuctionManager.java
@@ -1,11 +1,8 @@
 package it.unipi.dsmt.das.ejbs.beans.interfaces;
 
-import javax.ejb.Local;
 import javax.ejb.Remote;
 
 import it.unipi.dsmt.das.model.*;
-
-import java.util.List;
 
 
 @Remote
@@ -13,12 +10,12 @@ public interface AuctionManager {
 
 
     String createAuction(Auction auction);
-    String deleteAuction(int auctionId);
-    AuctionData selectAuction(int auctionId, int userId);
+    String deleteAuction(long auctionId);
+    AuctionData selectAuction(long auctionId, long userId);
     AuctionList auctionsList(int page);
-    AuctionList auctionAgentList(int agentId);
-    AuctionList auctionBidderList(int bidderId);
+    AuctionList auctionAgentList(long agentId);
+    AuctionList auctionBidderList(long bidderId);
 
     BidStatus makeBid(Bid bid);
-    BidStatus deleteBid(int auctionId, int bidId);
+    BidStatus deleteBid(long auctionId, long bidId);
 }
